Mount traversal prevention is a unix-only feature
