+++ pkg/extractor/dirwalk/platform_windows.go
@@ -1,0 +1,6 @@
+package dirwalk
+
+// return true if the dir is a different mount-point than its base
+func isDifferentMount(dir string) bool {
+	return false
+}
